tests/wpt/tests/css/css-tables/visibility-collapse-row-005.html

<!doctype html>
<meta charset="utf-8">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<link rel="author" title="Aleks Totic" href="mailto:atotic@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-tables-3/#computing-the-table-height">
<style>
  main table {
    border-spacing: 0;
  }
  main td {
    width: 50px;
    height: 50px;
    background: gray;
    padding: 0;
  }
</style>
<main>
  <h1>Visibility collapse and overflow</h1>
  <p>Collapsed row should not contribute to overflow</p>
  <div id="row" style="overflow-y: auto;  width: 200px; height: 100px;border:1px solid black;">
    <table>
      <tbody>
        <tr>
          <td></td>
          <td></td>
        </tr>
        <tr>
          <td></td>
          <td></td>
        </tr>
        <tr style="visibility: collapse">
          <td></td>
          <td></td>
        </tr>
      </tbody>
    </table>
  </div>
  <p>Collapsed section should not contribute to overflow</p>
  <div id="section" style="overflow-y: auto;  width: 200px; height: 50px;border:1px solid black;">
    <table>
      <thead>
        <tr>
          <td></td>
          <td></td>
        </tr>
      </thead>
      <tbody style="visibility:collapse">
        <tr>
          <td></td>
          <td></td>
        </tr>
        <tr>
          <td></td>
          <td></td>
        </tr>
      </tbody>
    </table>
  </div>
</main>
<script>
  test( _ => {
    assert_equals(document.querySelector("#row").scrollHeight, 100);
  }, "collapsed row should not contribute to overflow");
  test( _ => {
    assert_equals(document.querySelector("#section").scrollHeight, 50);
  }, "collapsed section should not contribute to overflow");
</script>
</html>
